Teach timestampConverter about PostgreSQL timestamptz type.
The timestampConverter forgets to list the PostgreSQL "timestamptz"
type, which stands for "timestamp with time zone". The already supported
"timestamp" datatype is in PostgreSQL a "timestamp without time zone", so
timestamp columns of type "timestamptz" won't be converted. This patch adds
the "timestamptz" type to the list of supported datatypes.