Introduce new configuration parameter format.timezone and fix format defaults.
The new configuration parameter "format.timezone" allows to select the
target time zone for conversions of timestamp into a specific timezone. The
default is "UTC", if omitted.

Formerly the timestamp converter used "YYYY" as a format pattern to convert
year notations in date, datetime and timestamp strings. This used to very
impractical, since in java.text.SimpleDateFormat this format string stands
for "Week Year" instead of the calendar year. Since "Week Year" is very
uncommon (and it is indeed contrary to "YYYY" in databases i know), we
change the default to "yyyy", which is, according to the Java API docs
the calendar year.