Make TimestampConverter smart for use with PostgreSQL.
Formerly, the TimestampConverter plugin used a super-flexible
approach with regular expression parsing the input string. This generic
approach works for simple temporal input strings, but not for everything.

Debezium for examples gets the decoded timestamp in UTC from PostgreSQL, but
the database doesn't necessarily pass all of the components of the time value.
It can happen that there's no "second" part of the timestamp, yielding in something
like this:

"2020-03-29T09:30Z"

The current generic mode is unable to parse this, since the used regex forces
seconds to be there. Fixing this would involve a much smarter regex, but i don't
think it is worth the trouble.

Instead, use a more strict approach which employs the (more recent) java.time
API and DateTimeFormatBuilder. This employs a flexible format pattern to parse the
input string into a java.time compatible object, allowing it to convert in any other
format we want. This makes it even possible to use time zone conversion via
ZonedDateTime, supporting the direct output of decoded values of type timestamptz
into a time zone supported by e.g. consumers.

We still keep the old behavior by splitting the legwork of the TimestampConverter
plugin into specific InternalConverter classes, GenericTimestampConverter and
PostgreSQLTimestampConverter. The first one is still the default, the latter
can be activated by specifying the property "convert.mode=postgresql".

In PostgreSQL conversion mode we support the PostgreSQL datatypes "time", "timetz",
"timestamp", "timestamptz" and "date".

Timezone conversion can only be done with values of type "timestamptz" and, if
explicitely activated, with "timetz". The latter is advised to be used with care,
see the updated README.md for details about this topic.

Unit tests included.