Introduce new output format parameter "format.datetimetz".
Formerly, if timezone conversion was enabled and someone wanted to preserve
the target timezone offset in the output value via "format.datetime" and
the corresponding format pattern, they would be in trouble if the same
configuration is used for "timestamp" and "timestamptz" column values.
In this specific case, it is not possible to specify offet patterns in the
format string, since the DateTimeFormatter would complain if a value
is converted into a type which doesn't support offsets, being the case
with "timestamp".

To solve this case, the configuration allows now to configure a separate
output format pattern for "timestamptz" values, "format.datetimetz", which
is used specificially for "timestamptz" output formatting.

Updated unit tests and documentation, as well as additional tests for
input formats.